Add a few more types to root
